content_files from the current day are ignored
Previously when the user decided to save the playlist content to the hard drive the
script wouldn't really work that day anymore. The newly saved content file would be
used in comparison to the current playlist content (which would be identical) and
the script would return there are no new songs.

Now the content_file from the current date will be ignored when comparing.