Gui can handle Spotify links for groups, playlists and artists
add method is_link(given_link) to URI_operations.py

The user can now enter a link as the target playlist when creating a group, or adding a playlist or
artist to a group. Previously this had to be a Spotify uri, but since Spotify made it more
complicated to get an uri, I added the possibility to enter links as well.