remove workflow to build on all apps
The building uses pytest to test the code, but I have not written any tests and probably wont't do so any time soon.
Also there is a weird bug that the windows test keeps failing because it requires a different syntax in the if-statement than the ubuntu and macos. I believe that fixing this issue will raise issues in the ubuntu and macos test.

TL;DR the automatic build was a nice gimmick, but right now its issues are bigger than its acual benefit. Maybe it will return in the future when I have written tests.